+++ cmd/geth/chaincmd.go
@@ -28,16 +28,6 @@ var (
 		Action: start,
 	}
 
-	importCmd = &cli.Command{
-		Name:   "import",
-		Usage:  "import missing blocks",
-		Action: restore,
-		Flags: []cli.Flag{
-			utils.SnapPath,
-			utils.SnapEngine,
-		},
-	}
-
 	chaindataCmd = &cli.Command{
 		Name:   "chaindata",
 		Usage:  "import missing blocks",
@@ -58,13 +48,13 @@ var (
 		},
 	}
 
-	batchimportcmd = &cli.Command{
-		Name:  "batchimport",
-		Usage: "batch import service",
+	importCmd = &cli.Command{
+		Name:  "import",
+		Usage: "import blocks service",
 		Flags: []cli.Flag{
 			utils.SvcHost,
 			utils.SvcPort,
 		},
-		Action: batchimport,
+		Action: importstore,
 	}
 )
